feat(): stores access data in a Redis instance intead of MySQL
Redis is good because fetching the data is fast and a TTL (Time To
Live) can be defined when setting a value.

Considering this, refresh tokens will not be stored more than 1 week,
without the need to manually flush them using a cron job or anything.
For the record, never deleted refresh tokens could happen if a client
never uses the refresh_token grant type to get a new access token.